add linter and resolve all issues
