--- src/about/AboutPage.js
@@ -1,61 +1,0 @@
-import { Container, Row} from "react-bootstrap";
-import './AboutPage.css';
-
-function AboutPage() {
-    return (
-        <div className="about">
-            <Container>
-                <Row>
-                    <h1 className="about-h1 first">Who We Are</h1>
-                    <p className="about-p">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do 
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. Odio 
-                        euismod lacinia at quis risus sed vulputate. Habitasse platea 
-                        dictumst vestibulum rhoncus est pellentesque elit ullamcorper dignissim. 
-                        Amet volutpat consequat mauris nunc congue nisi. Tristique risus nec 
-                        feugiat in fermentum posuere urna. Aliquet lectus proin nibh nisl. 
-                        Dui accumsan sit amet nulla facilisi morbi tempus. Feugiat in ante 
-                        metus dictum at tempor commodo ullamcorper a. Faucibus in ornare quam 
-                        viverra orci sagittis eu volutpat odio. Sapien faucibus et molestie ac 
-                        feugiat sed lectus vestibulum. Ultricies integer quis auctor elit. Non 
-                        blandit massa enim nec dui nunc mattis enim. Ullamcorper a lacus vestibulum 
-                        sed arcu non odio euismod lacinia. Viverra suspendisse potenti nullam 
-                        ac tortor vitae purus.</p>
-                </Row>
-                <Row>
-                    <h1 className="about-h1">What We Do</h1>
-                    <p className="about-p">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do 
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. Odio 
-                        euismod lacinia at quis risus sed vulputate. Habitasse platea 
-                        dictumst vestibulum rhoncus est pellentesque elit ullamcorper dignissim. 
-                        Amet volutpat consequat mauris nunc congue nisi. Tristique risus nec 
-                        feugiat in fermentum posuere urna. Aliquet lectus proin nibh nisl. 
-                        Dui accumsan sit amet nulla facilisi morbi tempus. Feugiat in ante 
-                        metus dictum at tempor commodo ullamcorper a. Faucibus in ornare quam 
-                        viverra orci sagittis eu volutpat odio. Sapien faucibus et molestie ac 
-                        feugiat sed lectus vestibulum. Ultricies integer quis auctor elit. Non 
-                        blandit massa enim nec dui nunc mattis enim. Ullamcorper a lacus vestibulum 
-                        sed arcu non odio euismod lacinia. Viverra suspendisse potenti nullam 
-                        ac tortor vitae purus.</p>
-                </Row>
-                <Row>
-                    <h1 className="about-h1">What We Stand For</h1>
-                    <p className="about-p last">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do 
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. Odio 
-                        euismod lacinia at quis risus sed vulputate. Habitasse platea 
-                        dictumst vestibulum rhoncus est pellentesque elit ullamcorper dignissim. 
-                        Amet volutpat consequat mauris nunc congue nisi. Tristique risus nec 
-                        feugiat in fermentum posuere urna. Aliquet lectus proin nibh nisl. 
-                        Dui accumsan sit amet nulla facilisi morbi tempus. Feugiat in ante 
-                        metus dictum at tempor commodo ullamcorper a. Faucibus in ornare quam 
-                        viverra orci sagittis eu volutpat odio. Sapien faucibus et molestie ac 
-                        feugiat sed lectus vestibulum. Ultricies integer quis auctor elit. Non 
-                        blandit massa enim nec dui nunc mattis enim. Ullamcorper a lacus vestibulum 
-                        sed arcu non odio euismod lacinia. Viverra suspendisse potenti nullam 
-                        ac tortor vitae purus.</p>
-                </Row>
-            </Container>
-        </div>
-    );
-}
-
-export default AboutPage;
