--- src/common/Footer.js
@@ -1,14 +1,0 @@
-import './Footer.css';
-
-function Footer() {
-    return (<footer className="mdl-mini-footer" id="footer">
-                <div className="mdl-mini-footer__left-section">
-                    <div className="mdl-logo">Copyright © Rochester Institute of Technology</div>
-                </div>
-                <div className="mdl-mini-footer__right-section">
-                    <div className="mdl-logo">One Lomb Memorial Drive Rochester NY 14623-5603</div>
-                </div>
-            </footer>);
-}
-
-export default Footer;
