--- src/pillars-temp/Mentoring.js
@@ -1,19 +1,0 @@
-
-import { Container, Row, Col } from 'react-bootstrap';
-import ProfileCard from '../common/ProfileCard';
-import './PillarsPage.css';
-
-function Mentoring() {
-    return(
-        <Container className="pillars">
-            <h1 className="pillar-h1">Mentoring</h1>
-            <p className="large-p">Mentoring Pillar things    
-            </p>
-            <Row>
-                <Col><ProfileCard name="Giovana Puccini"/></Col>
-            </Row>            
-        </Container>
-    );
-}
-
-export default Mentoring;
